requirements.txt
aiogram==3.6.0
python-dotenv==1.0.1
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0